move convertImage and tests to utils
